api-endpoints/requirements.txt
fastapi
google-cloud-firestore
uvicorn
pydantic
google-generativeai
fastmcp 
flask_cors
flask
google-cloud-storage
pydantic
python-multipart
firebase-admin
sentence-transformers
faiss-cpu
numpy
Pillow
python-dotenv
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
GoogleWalletPassGenerator
requests
openpyxl
google-genai
librosa
soundfile
scikit-learn
